+++ qt/gen_qabstracteventdispatcher.cpp
@@ -51,11 +51,11 @@ void QAbstractEventDispatcher_UnregisterSocketNotifier(QAbstractEventDispatcher*
 	self->unregisterSocketNotifier(notifier);
 }
 
-int QAbstractEventDispatcher_RegisterTimer(QAbstractEventDispatcher* self, int interval, uintptr_t timerType, QObject* object) {
+int QAbstractEventDispatcher_RegisterTimer(QAbstractEventDispatcher* self, int interval, int timerType, QObject* object) {
 	return self->registerTimer(static_cast<int>(interval), static_cast<Qt::TimerType>(timerType), object);
 }
 
-void QAbstractEventDispatcher_RegisterTimer2(QAbstractEventDispatcher* self, int timerId, int interval, uintptr_t timerType, QObject* object) {
+void QAbstractEventDispatcher_RegisterTimer2(QAbstractEventDispatcher* self, int timerId, int interval, int timerType, QObject* object) {
 	self->registerTimer(static_cast<int>(timerId), static_cast<int>(interval), static_cast<Qt::TimerType>(timerType), object);
 }
 
@@ -168,7 +168,7 @@ void QAbstractEventDispatcher_Delete(QAbstractEventDispatcher* self) {
 	delete self;
 }
 
-QAbstractEventDispatcher__TimerInfo* QAbstractEventDispatcher__TimerInfo_new(int id, int i, uintptr_t t) {
+QAbstractEventDispatcher__TimerInfo* QAbstractEventDispatcher__TimerInfo_new(int id, int i, int t) {
 	return new QAbstractEventDispatcher::TimerInfo(static_cast<int>(id), static_cast<int>(i), static_cast<Qt::TimerType>(t));
 }
