+++ qt/gen_qabstractitemdelegate.h
@@ -52,7 +52,7 @@ void QAbstractItemDelegate_SetEditorData(const QAbstractItemDelegate* self, QWid
 void QAbstractItemDelegate_SetModelData(const QAbstractItemDelegate* self, QWidget* editor, QAbstractItemModel* model, QModelIndex* index);
 void QAbstractItemDelegate_UpdateEditorGeometry(const QAbstractItemDelegate* self, QWidget* editor, QStyleOptionViewItem* option, QModelIndex* index);
 bool QAbstractItemDelegate_EditorEvent(QAbstractItemDelegate* self, QEvent* event, QAbstractItemModel* model, QStyleOptionViewItem* option, QModelIndex* index);
-struct miqt_string* QAbstractItemDelegate_ElidedText(QFontMetrics* fontMetrics, int width, uintptr_t mode, struct miqt_string* text);
+struct miqt_string* QAbstractItemDelegate_ElidedText(QFontMetrics* fontMetrics, int width, int mode, struct miqt_string* text);
 bool QAbstractItemDelegate_HelpEvent(QAbstractItemDelegate* self, QHelpEvent* event, QAbstractItemView* view, QStyleOptionViewItem* option, QModelIndex* index);
 struct miqt_array* QAbstractItemDelegate_PaintingRoles(const QAbstractItemDelegate* self);
 void QAbstractItemDelegate_CommitData(QAbstractItemDelegate* self, QWidget* editor);
@@ -65,7 +65,7 @@ struct miqt_string* QAbstractItemDelegate_Tr2(const char* s, const char* c);
 struct miqt_string* QAbstractItemDelegate_Tr3(const char* s, const char* c, int n);
 struct miqt_string* QAbstractItemDelegate_TrUtf82(const char* s, const char* c);
 struct miqt_string* QAbstractItemDelegate_TrUtf83(const char* s, const char* c, int n);
-void QAbstractItemDelegate_CloseEditor2(QAbstractItemDelegate* self, QWidget* editor, uintptr_t hint);
+void QAbstractItemDelegate_CloseEditor2(QAbstractItemDelegate* self, QWidget* editor, int hint);
 void QAbstractItemDelegate_connect_CloseEditor2(QAbstractItemDelegate* self, void* slot);
 void QAbstractItemDelegate_Delete(QAbstractItemDelegate* self);
